+++ src/app/dashboard/dashboard-routing.module.ts
@@ -1,3 +1,4 @@
+import { GeneratedListWrapperComponent } from './../generated-list-wrapper/generated-list-wrapper.component';
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { DashboardComponent } from './dashboard.component';
@@ -7,6 +8,10 @@ const routes: Routes = [
     { path: '',
       component: DashboardComponent,
       canActivate: [AuthGuard]
+    },
+
+    { path: 'list',
+      component: GeneratedListWrapperComponent,
     }
   ];
