+++ src/app/dashboard/dashboard.component.html
@@ -50,4 +50,3 @@
 
 </mat-vertical-stepper>
 
-<qpac-generated-list></qpac-generated-list>
