+++ src/app/generated-list-wrapper/generated-list-wrapper.component.html
@@ -1,0 +1,1 @@
+<qpac-generated-list></qpac-generated-list>
