+++ src/app/generated-list-wrapper/generated-list-wrapper.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'qpac-generated-list-wrapper',
+  templateUrl: './generated-list-wrapper.component.html',
+  styleUrls: ['./generated-list-wrapper.component.scss']
+})
+export class GeneratedListWrapperComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
