+++ src/app/generated-list-wrapper/generated-list/generated-list.component.html
