+++ src/app/generated-list-wrapper/pack-item/pack-item.component.ts
