--- src/app/generated-list/generated-list.component.scss
@@ -1,17 +1,0 @@
-.card {
-  padding: 50px;
-}
-.p {
-  padding: 15px;
-}
-.pack-items {
-  margin: 0;
-  padding: 20px;
-  li {
-    list-style: none;
-  }
-
-  li:not(:last-child) {
-    margin-bottom: 15px;
-  }
-}
