--- src/app/pack-item/pack-item.component.scss
@@ -1,11 +1,0 @@
-.pack-item {
-  background: #f0f8ff;
-  padding: 15px;
-  border-radius: 4px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  i {
-    cursor: pointer;
-  }
-}
